Agenda sale mas centrada ahora en la pantalla

+++ visita-facil-api/src/main/java/com/construsoft/visita_facil_api/repository/PropiedadRepository.java
@@ -1,12 +1,13 @@
 package com.construsoft.visita_facil_api.repository;
 
-import com.construsoft.visita_facil_api.model.Propiedad;
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
+import com.construsoft.visita_facil_api.model.Propiedad;
 
 @Repository
-public interface PropiedadRepository extends JpaRepository<Propiedad, Long> {
+public interface PropiedadRepository extends JpaRepository<Propiedad, Integer> {
     public List<Propiedad> findByDisponibleTrue();
 }
